feat(data_plane): create_data_plane encrypt fail
