+++ src/dashboard/apigateway/apigateway/service/gateway_jwt.py
@@ -15,35 +15,10 @@
 # We undertake not to change the open source license (MIT license) applicable
 # to the current version of the project delivered to anyone in the future.
 #
-from django.conf import settings
-from django.utils.encoding import force_bytes, smart_str
+from django.utils.encoding import smart_str
 
-from apigateway.common.encrypt.cipher import AESGCMCipher
 from apigateway.core.models import JWT, Gateway
-from apigateway.utils.crypto import BkCrypto, KeyGenerator
-
-
-class CustomCrypto:
-    def __init__(self):
-        self._jwt_cipher = AESGCMCipher(
-            force_bytes(settings.JWT_CRYPTO_KEY),
-            force_bytes(settings.CRYPTO_NONCE),
-        )
-
-    def encrypt(self, plaintext) -> str:
-        return self._jwt_cipher.encrypt_to_hex(plaintext)
-
-    def decrypt(self, encrypted_text) -> str:
-        return self._jwt_cipher.decrypt_from_hex(encrypted_text)
-
-
-def get_jwt_crypto():
-    if settings.BK_CRYPTO_TYPE == settings.CRYPTO_TYPE_APIGW_CUSTOM:
-        return CustomCrypto()
-    if settings.BK_CRYPTO_TYPE in ("SHANGMI", "CLASSIC"):
-        return BkCrypto()
-
-    raise ValueError(f"Unknown encrypt cipher type: {settings.BK_CRYPTO_TYPE}")
+from apigateway.utils.crypto import KeyGenerator, get_crypto
 
 
 class GatewayJWTHandler:
@@ -51,7 +26,7 @@ class GatewayJWTHandler:
     def create_jwt(gateway: Gateway) -> JWT:
         private_key, public_key = KeyGenerator().generate_rsa_key()
 
-        crypto = get_jwt_crypto()
+        crypto = get_crypto()
 
         return JWT.objects.create(
             gateway=gateway,
@@ -64,7 +39,7 @@ class GatewayJWTHandler:
 
     @staticmethod
     def update_jwt_key(gateway, private_key: bytes, public_key: bytes):
-        crypto = get_jwt_crypto()
+        crypto = get_crypto()
 
         jwt = JWT.objects.get(gateway=gateway)
         jwt.public_key = smart_str(public_key)
@@ -73,14 +48,14 @@ class GatewayJWTHandler:
 
     @staticmethod
     def get_private_key(gateway_id: int) -> str:
-        crypto = get_jwt_crypto()
+        crypto = get_crypto()
 
         jwt = JWT.objects.get(gateway_id=gateway_id)
         return crypto.decrypt(jwt.encrypted_private_key)
 
     @staticmethod
     def is_jwt_key_changed(gateway, private_key: bytes, public_key: bytes) -> bool:
-        crypto = get_jwt_crypto()
+        crypto = get_crypto()
 
         jwt = JWT.objects.get(gateway=gateway)
         return jwt.public_key != smart_str(public_key) or crypto.decrypt(jwt.encrypted_private_key) != smart_str(
