+++ src/dashboard/pyproject.toml
@@ -293,6 +293,7 @@ ignore_imports = [
     "apigateway.apps.data_plane.management.commands.* -> apigateway.controller.publisher.publish",
     "apigateway.apps.data_plane.management.commands.* -> apigateway.controller.tasks.*",
     "apigateway.apps.data_plane.management.commands.* -> apigateway.controller.constants",
+    "apigateway.apps.data_plane.management.commands.create_data_plane -> apigateway.controller.distributor.etcd",
     # -- celery task
     "apigateway.apps.permission.tasks -> apigateway.components.*",
     "apigateway.apps.permission.tasks -> apigateway.biz.permission",
@@ -358,6 +359,7 @@ ignore_imports = [
     "apigateway.apps.gateway.management.commands.export_inactive_gateways_by_stage -> apigateway.core.constants",
     "apigateway.apps.data_plane.management.commands.* -> apigateway.controller.publisher.*",
     "apigateway.apps.data_plane.management.commands.* -> apigateway.controller.tasks.*",
+    "apigateway.apps.data_plane.management.commands.create_data_plane -> apigateway.controller.distributor.etcd",
     "apigateway.apps.plugin.management.commands.export_plugin_binding_statistics -> apigateway.core.constants",
     # -- esb legacy
     "apigateway.apps.esb.*.* -> apigateway.core.models",
